refactor: clean up directory hierarchy and align namespaces
