+++ Editor/Assets/ColorPreset.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 65f37d9bfcf7495488efaeb2f79b4338
+guid: aba3a865aa9cf6a49a91d2baefd469b4
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
