+++ Editor/Assets/ColorPreset/DefaultColorPreset.asset.meta
