+++ Editor/Scripts/Inspector.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 2ff943d0d05be214f9eeff111dfff990
+guid: d2037d139a3970a4b95273ad965fb81e
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
