+++ EditorOnly/Scripts/Components/ColorManager.cs.meta
