+++ EditorOnly/Scripts/dev.limitex.mono-ui.editor-only.asmdef
@@ -1,5 +1,5 @@
 {
-    "name": "dev.limitex.mono-ui.udon.scripts",
+    "name": "dev.limitex.mono-ui.editor-only",
     "rootNamespace": "",
     "references": [
         "GUID:6055be8ebefd69e48b49212b09b47b2f"
