+++ Runtime/Udon/UdonSharp.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 147f41f3108cda84f82767cddb8a416c
+guid: c75ee192d475bd64381b32d8dd91291f
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
